introduce-session-controller(move): Hand session lifecycle over to SessionController
Pure physical move per MECH_COMMIT_SPLIT §"split-class scenario". Cut
@staticmethod open_session + close_session from TokenizerControlMixin and
@staticmethod handle_open_session_req_output from TokenizerManager;
paste into SessionController (drop @staticmethod, replace
``self: "SessionController"`` → plain ``self``). Entrypoint callers
(engine.py, http_server.py) get pure prefix replacement:
``TokenizerManager.<method>(...session_controller, ...)`` →
``...session_controller.<method>(...)``. The request-dispatcher entry for
OpenSessionReqOutput flips from the prep-stage lambda forwarder to a direct
bound method ``self.session_controller.handle_open_session_req_output``, and
the now-unused SessionController TYPE_CHECKING import is removed from
tokenizer_control_mixin.py.

Refactor chain ID: introduce-session-controller-move
PR-Title: Extract SessionController from TokenizerManager
