+++ python/sglang/srt/managers/tokenizer_manager.py
@@ -431,9 +431,7 @@ class TokenizerManager(TokenizerControlMixin):
                 (AbortReq, self._handle_abort_req),
                 (
                     OpenSessionReqOutput,
-                    lambda recv_obj: TokenizerManager.handle_open_session_req_output(
-                        self.session_controller, recv_obj
-                    ),
+                    self.session_controller.handle_open_session_req_output,
                 ),
                 (
                     UpdateWeightFromDiskReqOutput,
@@ -1431,18 +1429,6 @@ class TokenizerManager(TokenizerControlMixin):
 
     def update_active_ranks(self, ranks: ActiveRanksOutput):
         self.send_to_scheduler.send_pyobj(ranks)
-
-    @staticmethod
-    def handle_open_session_req_output(self: "SessionController", recv_obj):
-        future = self.session_futures.get(recv_obj.session_id)
-        if future is None:
-            logger.warning(
-                "Open session response arrived after waiter cleanup: %s",
-                recv_obj.session_id,
-            )
-            return
-        if not future.done():
-            future.set_result(recv_obj.session_id if recv_obj.success else None)
 
     def _handle_update_weights_from_disk_req_output(self, recv_obj):
         if self.server_args.dp_size == 1:
